Changes thread idx jobs to be executed on the same thread. Very very WIP and hacky tho.

+++ include/job_queue.h
@@ -24,7 +24,7 @@ private:
     //        type automatically loop as an increment overflows it.
     // I think `uint16_t` will be the largest allowable size for a looping type
     // ngl.  -Thea 2024/12/14
-    using looping_numeric_t = uint16_t;
+    using looping_numeric_t = uint8_t;
     inline static const size_t k_pointer_buffer_indices{ std::numeric_limits<looping_numeric_t>::max() + 1 };
     std::atomic<void*> m_pointer_buffer[k_pointer_buffer_indices];
     std::atomic<looping_numeric_t> m_front_idx{ 0 };  // For reserving front buffer ptr.
